Added integer enlarging. Doesn't work properly

+++ lab4/lab4.cpp
@@ -8,7 +8,7 @@ int main() {
     Image image;
     string name1 = "test.bmp";
     string name2 = "new.bmp";
-    double ext = 2.0;
+    int ext = 2;
     image.ResizeImage(name1, name2, ext);
     system("pause");
     return 0;
